Addressed martinvw's comments.
Includes:
- transformation service moved;
- some missing lines at the end of the files added;
- other minor imporvements.
