--- addons/binding/org.openhab.binding.feed.transform.test/src/test/resources/input/rss_2.0_empty.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<rss>
-</rss>
