--- addons/binding/org.openhab.binding.feed.transform/META-INF/MANIFEST.MF
@@ -1,16 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Feed Transformation Service
-Bundle-SymbolicName: org.openhab.binding.feed.transform
-Bundle-Vendor: openHAB
-Bundle-Version: 2.1.0.qualifier
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ClassPath: .
-Import-Package: com.google.common.collect,
- com.rometools.rome.feed.synd,
- com.rometools.rome.io,
- org.apache.commons.io,
- org.eclipse.smarthome.core.transform,
- org.slf4j
-Bundle-ActivationPolicy: lazy
-Export-Package: org.openhab.binding.feed.transform
