+++ addons/transform/org.openhab.transform.feed.test/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.openhab.binding.feed.transform.test</name>
+	<name>org.openhab.transform.feed.test</name>
 	<comment>Feed Transformation Service Tests</comment>
 	<projects>
 	</projects>
