+++ addons/transform/org.openhab.transform.feed.test/build.properties
