+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/input/rss_2.0_empty.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<rss>
+</rss>
