+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/output/atom_0.3.xml
