+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/output/atom_1.0.xml
