+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/output/rss_0.92.xml
