+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/output/rss_1.0.xml
