+++ addons/transform/org.openhab.transform.feed.test/src/test/resources/output/rss_2.0.xml
