+++ addons/transform/org.openhab.transform.feed/OSGI-INF/feedTransformationService.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+  Copyright (c) 2010-2017 by the respective copyright holders.
+ 
+  All rights reserved. This program and the accompanying materials
+  are made available under the terms of the Eclipse Public License v1.0
+  which accompanies this distribution, and is available at
+  http://www.eclipse.org/legal/epl-v10.html
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="feedTransformationService">
+   <implementation class="org.openhab.transform.feed.FeedTransformationService"/>
+   
+   <service>
+      <provide interface="org.eclipse.smarthome.core.transform.TransformationService"/>
+   </service>
+   <property name="smarthome.transform" type="String" value="FEED"/>
+</scr:component>
