+++ addons/transform/org.openhab.transform.feed/build.properties
