+++ addons/transform/pom.xml
@@ -2,25 +2,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
   <parent>
-    <groupId>org.openhab</groupId>
-    <artifactId>pom-addons2</artifactId>
+    <groupId>org.openhab.addons</groupId>
+    <artifactId>pom</artifactId>
     <version>2.1.0-SNAPSHOT</version>
   </parent>
 
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.openhab.addons</groupId>
+  <groupId>org.openhab.transform</groupId>
   <artifactId>pom</artifactId>
 
-  <name>openHAB Add-ons</name>
+  <name>openHAB Transformation Services</name>
 
   <packaging>pom</packaging>
 
   <modules>
-    <module>binding</module>
-    <module>io</module>
-    <module>transform</module>
-    <module>ui</module>
-    <module>voice</module>
+    <module>org.openhab.transform.feed</module>
+    <module>org.openhab.transform.feed.test</module>
   </modules>
 
 </project>
