+++ features/openhab-addons/src/main/feature/feature.xml
@@ -363,10 +363,10 @@
         <bundle start-level="80">mvn:org.openhab.io/org.openhab.io.imperihome/${project.version}</bundle>
     </feature>
 
-    <feature name="openhab-misc-feed" description="Feed Transformation Service" version="${project.version}">
+    <feature name="openhab-transform-feed" description="Feed Transformation Service" version="${project.version}">
         <feature>openhab-runtime-base</feature>
         <feature>openhab-transport-feed</feature>
-        <bundle start-level="80">mvn:org.openhab.binding/org.openhab.binding.feed.transform/${project.version}</bundle>
+        <bundle start-level="80">mvn:org.openhab.binding/org.openhab.transform.feed/${project.version}</bundle>
     </feature>
 
     <!-- voice -->
